I have created and trained a lightGBM model to predict Bryce Harper's wOBA for the next 3 seasons (2025, 2026, 2027). I started by loading in all batting stats for all players from 2015-2024. I then began creating my future target (next year's wOBA). I did this by renaming columns using key value pairs to make it easier to access these columns in the future. I then selected my features and defined my target to begin training the model. Once the model was fit, I found the mean squared error to be 0.0010 which is great! The low value demonstrates this model's accuracy. I then filtered the data to get Bryce Harper's specific statistical information. Based on this model, I predict Harper's wOBA to be .378 for the 2025 season.
I then expanded on this model and made predictions for the next 3 seasons . I used the same model but now targeting the next 3 years instead of just 2025. I used the same features as well. Based on this model, I predict Harper's wOBA for 2025, 2026 and 2027 to be .345, .369, and .328 respectively. For reference, FanGraphs currently predicts Harper's wOBA to be .379, .362, and .315 for 2025, 2026 and 2027 respectively. I feel that this model is very accurate given the low mean squared error values as well as FanGraphs having very similar predictions.