I created a model to build a my version of a specific scouting report for Phillies hitters vs Shohei Ohtani. The scouting report will look at the Phillies hitters xwOBA values based on certain pitch types Ohtani throws. Ohtani got the best of the Phillies line up the first, and only time he has faced them in his career. On 9/16/2025, Ohtani threw 68 pitches without surrendering a hit. Dominating using his four-seam fastball, throwing it 51% of the time. Ohtani has 7 pitches in his arsenal and he mixed his pitches well agains the Phillies. 35 FF, 12 SL, 7 CU, 6 ST, 4 SI, 3 FC and 1 FS. He touched 101mph on his fastball that outing as well. Only 4 balls were hard hit by the Phillies that night, 3 FF and 1 FC, all line drive outs. Given the Phillies will be seeing Ohtani to start the postseason, I wanted to look deeper into his arsenal and find where the Phillies may be able to capitalize on him.
Ohtani has a 1.71 ERA at home in 2025, which is where he dominated the Phillies' lineup. However, he has a 3.81 ERA on the road. It will be interesting to see if Citizens Bank Park's playoff atmosphere will add to that total as well. The Phillies threw a bunch of left handed hitters at Ohtani that night he dominated them at his home ballpark. Typically, Ohtani will throw mainly FF, SL, CU to left handed hitters but mixes in his FC, ST, FS and SI as needed. He was able to blow the fastball by the hitters for either swing and miss or a ground ball/ fly out. If the hitter fouled off the fastball he was able to get swing and miss with the slider and curve. Lefties typically hit the cutter and sinker the best off of Ohtani. As a whole the Phillies hit the fastball well, but they have a tendency, despite that number dropping this year, to chase pitches out of the zone. Their chase rate is about 2% higher than MLB average in 2025.

Now for right handed hitters, Ohtani will mainly throw his ST, FF, SI, SL while mixing in his CU, FC and FS when needed. Hitters appear to have the best chance against the SI, hitting .571 and the FC if its thrown. They key is going to be able handling the fastball when its pitched in the zone. Max Kepler specifically has excellent numbers against the fastball, being able to hit high velocity very well. He had a few pitches to hit last he saw Ohtani, timing will be crucial. Staying within the zone will be crucial as well. The Phillies cannot expand the zone looking to make something happen. Ohtani is very good at putting pitches on the corners, but will come over the middle of the plate. Being patient will be key.

In his 47 IP this season, Ohtani has only allowed 3% of his pitches to be barreled which is incredible. The Phillies should look to foul some pitches off and create longer at-bats since the most pitches he has thrown all season is 92 in 6 innings in his last start of the season against Arizona. It will be a tough outing against an elite pitcher, but making sure they do not chase, and being able to create a high pitch count will prove to be beneficial. Their timing on fastballs over the middle of the plate will be in focus as well if they have success. 

The Phillies were without Alec Bohm, Trea Turner and Edmundo Sosa during this outing as well. Those Phillies have obviously not faced Ohtani in their careers. Turner and Bohm, two likely game 1 starters, have great xwOBA values on the FF. It will be great to see how the team matches up in a huge game for both sides. Looking forward to seeing how they match up this time around!