This model is identical to the model I created on the Phillies vs Ohtani. I wanted to use the same format to analyze Cristopher Sánchez's pitch arsenal against the Dodgers Hitters xwOBA values. Sánchez has 3 (CH, SL, SI) pitches in his arsenal which, today may not be typical for a starting pitcher, but he pitches with such power using the pitches he has. His most dominant pitch is his changeup which he can throw to lefties and righties for swing and miss. Sánchez had the best year of his career in 2025 reaching 200 strikeouts for the first time and surpassing 200 IP. Looking back on the 2 games Sánchez threw against the Dodgers this year, in comparison to the results created using this model, I have drawn several conclusions. On 4/6/2025 Sánchez threw 88 pitches in 5.2 innings allowing 4 ER. He used his SI 56% of the time in this outing while mixing in his SL 23% of the time and his CH 22% of the time. Overwhelmingly, the sinker was hit the hardest and the Dodgers had the most success with it. The sinker is a good pitch for Sánchez but he seemed to be throwing it middle-middle to an excellent lineup.
In his second outing against the Dodgers on 9/16, he also surrendered 4 ER. However, the damage was done early as Sánchez managed to complete 7.0 IP on 102 pitches. He again used his SI the most at 61%, he did utilize his CH slightly more at 24% and his slider slightly less at 16%. Again, the majority of pitches hit for damage were done on the sinker staying in the middle of the plate and elevated. 

Between both starts the Dodger with the most success against Sánchez in 2025 is Teoscar Hernández going 3-6 with 2 homers. Both of those homers on middle-middle sinkers. The majority of the strikeouts Sánchez has against the Dodgers is on the changeup from both outings. Sánchez has such an excellent changeup he should utilize it more to this great lineup. They are looking to put the ball in play, so jamming the hitter inside using the sinker to induce ground balls and throwing the changeup in deep counts for swing and miss will prove to be successful. Sánchez has to make sure he has great command of the sinker on the inner part of the plate and he will have success. 

Pitching backward, meaning throwing secondary pitches like the changeup early in counts may also be an alternative plan to keep the hitters guessing. Again, the Dodgers are eager to put the ball in play and swing early. Getting swing and miss right away with the changeup for strike one can allow for a sinker to freeze the hitter later in the at bat. This may be a good adjustment strategy for the second or third time through the order. 

I look forward to seeing this matchup and the strategy of both starting pitchers. (.324 for average xwOBA is the MLB 2025 average).